refactor: isolate remote mutation coordination
