+++ src/features/import/hooks/useDeckImport.ts
@@ -5,8 +5,8 @@ import * as action from "@/action";
 import { useAuth } from "@/auth/AuthContext";
 import { useCardMutations } from "@/features/card/hooks/useCardMutations";
 import { useDeckMutations } from "@/features/deck/hooks/useDeckMutations";
+import { CardBulkMutationError } from "@/query/mutations/cardMutationService";
 import { useRemoteCollections } from "@/query/useRemoteCollections";
-import { CardBulkMutationError } from "@/query/cardMutationService";
 import { useConfig } from "@/hooks/useConfig";
 import type { DeckImportPreview, DeckImportResult, DeckImportRow } from "@/features/import/components/deckImportTypes";
 import { buildDeckImportPlan, parseDeckImportCsv } from "@/features/import/lib/deckImportAnalysis";
